% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai.R
\name{format_function_call_df}
\alias{format_function_call_df}
\title{Format the function call dataframe. Nest parameters based on the group column and format the list structure for conversion to JSON}
\usage{
format_function_call_df(function_call_df)
}
\arguments{
\item{function_call_df}{}
}
\value{
A formatted tibble
}
\description{
Format the function call dataframe. Nest parameters based on the group column and format the list structure for conversion to JSON
}
\examples{
function_call_df <- get_function_call_params() |> format_function_call_df()
}
