% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai.R
\name{format_openai_results}
\alias{format_openai_results}
\title{Convert the JSON formatted openAI responses into a more human readable nested tibble.}
\usage{
format_openai_results(model_response)
}
\arguments{
\item{model_response}{}
}
\value{
A nested tibble
}
\description{
Convert the JSON formatted openAI responses into a more human readable nested tibble.
}
\examples{
#' specific_context <- "Chronic wasting disease (CWD) is a fatal, prion disease of cervids that was first detected in Alberta in 2005"
common_context <- "Please use the extract_outbreak_data function to extract outbreak details from the given abstract and return only the results"
chat_results <- openai_chat(specific_context, common_context)
formatted_chat_results <- format_openai_results(chat_results)
}
