% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai.R
\name{get_function_call_parameter}
\alias{get_function_call_parameter}
\title{Arrange parameter details in JSON schema format}
\usage{
get_function_call_parameter(
  parameter_name,
  description,
  type = c("string", "null"),
  enum = NULL
)
}
\arguments{
\item{parameter_name}{The name of the parameter openAI should return}

\item{description}{A description of the parameter. OpenAI will use this description to construct it's response}

\item{type}{What type or types are allowed in OpenAI's response. i.e. c("string", "null") constrains openAI to returning a character string or NULL.}

\item{enum}{A list of the return values that OpenAI should adhere to when choosing a response.}
}
\value{
A formatted and named parameter list
}
\description{
Arrange parameter details in JSON schema format
}
\examples{
function_call_parameter <- get_function_call_parameter("disease", "The outbreak event disease name")
}
