% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai.R
\name{get_function_call_params}
\alias{get_function_call_params}
\title{Construct a tibble of parameters that openAI will return and specify constraints such as type and enumeration}
\usage{
get_function_call_params(country = NA, disease = NA)
}
\arguments{
\item{country}{Country context can be added to parameter descriptions by modifying the tribble to include glue or paste statements}

\item{disease}{Disease context can be added to parameter descriptions by modifying the tribble to include glue or paste statements}
}
\value{
A tibble outlining the parameters that openAI should use in it's reply. Will be further processed by format_function_call_df
}
\description{
Construct a tibble of parameters that openAI will return and specify constraints such as type and enumeration
}
\examples{
function_call_params <- get_function_call_params()
}
