% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai.R
\name{get_openai_function_call}
\alias{get_openai_function_call}
\title{context aware openai function call.}
\usage{
get_openai_function_call(
  country = NA,
  disease = NA,
  function_name = "exract_outbreak_data",
  function_description = "Extract outbreak details from provided abstract"
)
}
\arguments{
\item{country}{Passed to get_function_call_params if parameter descriptions need to include context}

\item{disease}{Passed to get_function_call_params if parameter descriptions need to include context}

\item{function_name}{The name of the imaginary function openAI will be formatting it's response for. Should be descriptive of the desired task it will perform.}

\item{function_description}{# A description of the what the function would accomplish}
}
\value{
Returns a list containing a formatted function call as an R list, as a compact JSON, a pretty JSON, and a tibble of the parameters used in constructing the call.
}
\description{
context aware openai function call.
}
\examples{
function_call <- get_openai_function_call()
}
