% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai.R
\name{openai_chat}
\alias{openai_chat}
\title{Initiate a conversation with openAI using the provided model and system, common, and specific context. Note: to use these functions you need to provide an openai key in the .env file as in:}
\usage{
openai_chat(
  specific_context,
  common_context,
  hint = NA,
  model = "gpt-3.5-turbo",
  system_context =
    "You act as a function to extract outbreak infromation from a provided abstract."
)
}
\arguments{
\item{specific_context}{This is the text specific to each query you would like OpenAI to consider}

\item{common_context}{This is the text OpenAI should consider in every query}

\item{hint}{An optional hint to get OpenAI on the right track.}

\item{model}{The identity of the model you would like to submit the context to.}

\item{system_context}{Text explaining to the model what role you would like it to play. In example, "Respond as if you an 18th century naturalist."}
}
\value{
A tibble containing the responses from openAI for each query
}
\description{
Initiate a conversation with openAI using the provided model and system, common, and specific context. Note: to use these functions you need to provide an openai key in the .env file as in:
}
\examples{
specific_context <- "Chronic wasting disease (CWD) is a fatal, prion disease of cervids that was first detected in Alberta in 2005"
common_context <- "Please use the extract_outbreak_data function to extract outbreak details from the given abstract and return only the results"
chat_results <- openai_chat(specific_context, common_context)
}
